WIP : Improving docstrings
